레시피 후기 페이지 수정중

+++ src/main/java/com/project/foodsite/dao/ReviewDAO.java
@@ -27,9 +27,9 @@ public interface ReviewDAO {
 
     int reviewInsert(ReviewVO review);
 
-    ReviewDetailDTO selectreview(long review_id);
+    ReviewDetailDTO selectreview(int review_id);
 
-    int deletereview(long review_id);
+    int deletereview(int review_id);
 
 
 }
